requirements.txt
Django
django-star-ratings
pillow
django-ipware
